Changes over the Dijkstra Algorithm

+++ src/decks/City.java
@@ -12,9 +12,9 @@ package decks;
 public class City {
     
     private String Name; //Name to show in the Deck of cards
-    private float Longitude; //Parameters to pass to the api call
-    private float Latitude; //Parameters to pass to the api call
-    private float Radio; //Parameters to pass to the api call
+    private double Longitude; //Parameters to pass to the api call
+    private double Latitude; //Parameters to pass to the api call
+    private double Radio; //Parameters to pass to the api call
 
     /**
      * Builder Method for the City Class
@@ -39,30 +39,27 @@ public class City {
         this.Name = Name;
     }
 
-    public float getLongitude() {
+    public double getLongitude() {
         return Longitude;
     }
 
-    public void setLongitude(float Longitude) {
+    public void setLongitude(double Longitude) {
         this.Longitude = Longitude;
     }
 
-    public float getLatitude() {
+    public double getLatitude() {
         return Latitude;
     }
 
-    public void setLatitude(float Latitude) {
+    public void setLatitude(double Latitude) {
         this.Latitude = Latitude;
     }
 
-    public float getRadio() {
+    public double getRadio() {
         return Radio;
     }
 
-    public void setRadio(float Radio) {
+    public void setRadio(double Radio) {
         this.Radio = Radio;
     }
-    
-    
-    
 }
